introduce JsonProvider
Requiring PyYAML is kind of annoying, so we're introducing
the ability to use JSON files to configuration management.

This allows YAMZ to be used in environments where there are size constraints, such as AWS Lambda.
Also, some folks prefer not to work with YAML, so this, should, in theory, make them less cranky 🙃